docs: improved documentation (and minor tidying)
