+++ src/utils/__test__/slugify.test.ts
@@ -1,8 +1,6 @@
 import assert from 'node:assert';
 import { describe, it } from 'node:test';
 
-import { pathToFileURL } from 'url';
-
 import { slugify } from '../slugify.ts';
 
 const slugChars = 'abcdefghijklmnopqrstuvwxyz0123456789_.-';
@@ -25,9 +23,6 @@ describe('slugify', () => {
     assert.equal(slugify("Hawke's Bay"), 'hawkes-bay');
   });
   it('should replace slashes with hyphens', () => {
-    const x = 'Tikitapu/Blue Lake';
-    const u = pathToFileURL(x);
-    console.log('url', u);
     assert.equal(slugify('Tikitapu/Blue Lake'), 'tikitapu-blue-lake');
   });
   it('should replace commas with hyphens', () => {
